requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
boto3==1.42.62
botocore==1.42.62
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.4
cron-descriptor==1.4.5
Django==6.0.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-extensions==4.1
django-ipware==7.0.1
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.8.0
flower==2.0.1
frozenlist==1.8.0
geoip2==5.2.0
humanize==4.15.0
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
kombu==5.6.2
maxminddb==3.1.1
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
pytz==2026.1.post1
redis==7.3.0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==202603
urllib3==2.6.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.6.0
yarl==1.23.0
